export UV_PROJECT_ENVIRONMENT and UV_PYTHON with virtualenv information (#8)
This patch adds the session support for UV_PROJECT_ENVIRONMENT and UV_PYTHON to
be forwarded to ALL run commands in the session after it has been posted. It
limits exposure of environment variables to only projects with-in the current
session so that the we don't pollute the os.environ process wide environment.